impl 0.01 STEP 1

+++ src/main/MyPoint_H_M.java
@@ -18,7 +18,7 @@ public class MyPoint_H_M extends MyPoint {
     private double rangeStart;
     private double rangeEnd;
 
-    public MyPoint_H_M(int y, double y_, double rangeStart, double rangeEnd) {
+    public MyPoint_H_M(double y, double y_, double rangeStart, double rangeEnd) {
         super(y, y_);
         this.rangeStart = rangeStart;
         this.rangeEnd = rangeEnd;
